# api-gateway/docs/swagger.yaml
...
openapi: 3.0.3
info:
  title: Granula API
  version: 1.0.0
  description: |
    # Granula API

    API для интеллектуального сервиса планирования ремонта и перепланировки квартир.

    ## Возможности

    - 🏠 **Загрузка планировок** - Загружайте изображения планировок для автоматического распознавания
    - 🤖 **AI распознавание** - Автоматическое распознавание комнат, стен и элементов
    - 🎨 **3D редактор** - Редактирование планировки в интерактивном 3D редакторе
    - ✅ **Проверка норм** - Автоматическая проверка на соответствие СНиП и ЖК РФ
    - 💡 **AI генерация** - Генерация вариантов планировки с помощью AI
    - 📋 **Заявки экспертам** - Подача заявок на консультации специалистов БТИ

    ## Аутентификация

    API использует JWT (JSON Web Tokens) для аутентификации.

    ### Получение токенов

    1. Зарегистрируйтесь через `POST /api/v1/auth/register`
    2. Или войдите через `POST /api/v1/auth/login`
    3. Получите `access_token` и `refresh_token`

    ### Использование токенов

    Передавайте Access Token в заголовке Authorization:
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Обновление токенов

    Access Token действует 15 минут. Для обновления используйте `POST /api/v1/auth/refresh`

    ## Rate Limiting

    | Тип запроса | Лимит |
    |-------------|-------|
    | Анонимные | 60 req/min |
    | Аутентифицированные | 300 req/min |
    | AI endpoints | 30 req/min |

    При превышении лимита возвращается HTTP 429.

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен аутентификации. Получите через `/auth/login` или `/auth/register`.

        Формат: `Bearer <token>`

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: VALIDATION_ERROR
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: Неверный формат email
        details:
          type: object
          additionalProperties:
            type: string
          description: Детали ошибок валидации по полям
          example:
            email: "Неверный формат email"
            password: "Минимум 8 символов"
        request_id:
          type: string
          description: ID запроса для отладки
          example: req-abc123def456
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Общее количество элементов
          example: 150
        page:
          type: integer
          description: Текущая страница
          example: 1
        page_size:
          type: integer
          description: Элементов на странице
          example: 20
        total_pages:
          type: integer
          description: Всего страниц
          example: 8
        has_next:
          type: boolean
          description: Есть следующая страница
          example: true
        has_prev:
          type: boolean
          description: Есть предыдущая страница
          example: false

    # -------------------------------------------------------------------------
    # Auth Schemas
    # -------------------------------------------------------------------------
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID воркспейса
          example: 550e8400-e29b-41d4-a716-446655440000
        owner_id:
          type: string
          format: uuid
          description: UUID владельца
          example: 660e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          description: Название
          example: Перепланировка квартиры на Тверской
        description:
          type: string
          description: Описание
          example: Проект перепланировки 3-комнатной квартиры
        member_count:
          type: integer
          description: Количество участников
          example: 3
        project_count:
          type: integer
          description: Количество проектов
          example: 2
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-11-29T14:20:00Z"
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
    CreateWorkspaceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Название воркспейса
          example: Перепланировка квартиры
        description:
          type: string
          maxLength: 1000
          description: Описание
          example: Проект перепланировки 3-комнатной квартиры
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 770e8400-e29b-41d4-a716-446655440002
        user_id:
          type: string
          format: uuid
          example: 880e8400-e29b-41d4-a716-446655440003
        role:
          type: string
          enum: [owner, admin, editor, viewer]
          description: |
            Роль участника:
            - **owner** - Полный контроль, может удалить воркспейс
            - **admin** - Управление участниками и настройками
            - **editor** - Редактирование контента
            - **viewer** - Только просмотр
          example: editor
        joined_at:
          type: string
          format: date-time
          example: "2024-02-20T12:00:00Z"
    AddMemberRequest:
      type: object
      required:
        - user_id
        - role
      properties:
        user_id:
          type: string
          format: uuid
          description: UUID пользователя
          example: 880e8400-e29b-41d4-a716-446655440003
        role:
          type: string
          enum: [admin, editor, viewer]
          description: Роль (owner назначается автоматически)
          example: editor

    # -------------------------------------------------------------------------
    # Floor Plan Schemas
    # -------------------------------------------------------------------------
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /workspaces:
    get:
      tags: [workspaces]
      summary: Список воркспейсов
      description: Возвращает список воркспейсов текущего пользователя
      operationId: listWorkspaces
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: name
          in: query
          description: Фильтр по названию
          schema:
            type: string
      responses:
        '200':
          description: Список воркспейсов
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [workspaces]
      summary: Создать воркспейс
      operationId: createWorkspace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Воркспейс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /workspaces/{id}:
    get:
      tags: [workspaces]
      summary: Получить воркспейс
      operationId: getWorkspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Данные воркспейса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          description: Не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [workspaces]
      summary: Обновить воркспейс
      operationId: updateWorkspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '200':
          description: Воркспейс обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    delete:
      tags: [workspaces]
      summary: Удалить воркспейс
      description: Удаляет воркспейс (только владелец)
      operationId: deleteWorkspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Воркспейс удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /workspaces/{id}/members:
    get:
      tags: [workspaces]
      summary: Список участников
      operationId: getWorkspaceMembers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
    post:
      tags: [workspaces]
      summary: Добавить участника
      operationId: addWorkspaceMember
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Участник добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'

  # ---------------------------------------------------------------------------
  # Floor Plan Endpoints
  # ---------------------------------------------------------------------------
...